# Improve list conversation performance