We're seeing the app freeze frequently for users. On export it appears there is some performance issues with importing and listing conversations

Some stats

> 1:51:21 PM | DEBUG : [XmtpRN] Imported 846 exported conversations into client in 66.073s
> 1:51:36 PM | DEBUG : [XmtpRN] Listing 0 conversations for 0x2659aBe3F33c3BA20B9F440f24b746A22b1d53F7 took 14.756 seconds
> 1:51:37 PM | DEBUG : [XmtpRN] Listing 846 conversations for 0xf0EA7663233F99D0c12370671abBb6Cca980a490 took 81.935 seconds

This parallelizes some stuff during import and converts known topics to a set to help with performance.

The SDK lists 2k conversations in ~ 10s so 14 seconds to load 0 conversations looks like potentially an issue in this code somewhere.