@cameronvoell 2024-08-30
curious - what device did you capture those logs on? I just tried on my Android Pixel and got (on `main` branch just for comparison sake)
```
[XmtpRN] Listing 1995 conversations for 0x209fAEc92D9B072f3E03d6115002d6652ef563cd took 28.53 seconds 
...
[XmtpRN] Listing 0 conversations for 0x94f3099AE62E5dA07113016144aef87929683bEf took 0.414 seconds 
```