@nplasterer 2024-08-30
@cameronvoell these logs are from @saulmc device when he saw the device frozen. Logs in slack.